This project's main goal is to provide an automated and online  billing platform for electricity consumers
his Simple Electricity Billing System Project is accessible to both the management and the clients.
It has 2 different user roles which are the Administrator and the Client. The Clients are required to register
the system which requires them to fill in some relevant data such as their name, billing address, and connection ID.